Created Home Page (Background image, moving greetings, and text-mirroring)
Started working on the home page. It now has each word in "Hello!" move when cursor is hovering over it. The general opener line has a reflection and will stay as one line across the screen, even with smaller screen size.
TODO: Add icons that will navigate to another page when clicked on.